uodated some files to match code organization from meters, conversions, etc. Moved files inside users folder, removed old files since we are now using modals.
Still needs the Save User Edits addressed as it's currently throwing an error.
