+++ src/client/app/translations/data.ts
@@ -162,6 +162,7 @@ const LocaleTranslationData = {
 		"edit.group": "Edit Group",
 		"edit.meter": "Edit Meter",
 		"edit.unit": "Edit Unit",
+		"edit.user": "Edit User",
 		"email": "Email",
 		"enable": "Enable",
 		"error.bar": "Show error bars",
@@ -480,11 +481,12 @@ const LocaleTranslationData = {
 		"upload.readings.csv": "Upload readings CSV file",
 		"used.so.far": "used so far",
 		"used.this.time": "used this time",
+		"user.confirm.delete": "Delete the user ",
+		"user.password.mismatch": "Passwords Do Not Match",
 		"users": "Users",
 		"users.failed.to.create.user": "Failed to create a user.",
 		"users.failed.to.delete.user": "Failed to delete a user.",
 		"users.failed.to.edit.users": "Failed to edit users.",
-		"user.password.mismatch": "Passwords Do Not Match",
 		"users.successfully.create.user": "Successfully created a user.",
 		"users.successfully.delete.user": "Successfully deleted user.",
 		"users.successfully.edit.users": "Successfully edited users.",
@@ -653,6 +655,7 @@ const LocaleTranslationData = {
 		"edit.group": "Modifier Groupe",
 		"edit.meter": "Modifier Métre",
 		"edit.unit": "Edit Unit\u{26A1}",
+		"edit.user": "Modifier l'utilisateur",
 		"email": "E-mail",
 		"enable": "Activer",
 		"error.bar": "Show error bars\u{26A1}",
@@ -972,6 +975,7 @@ const LocaleTranslationData = {
 		"used.so.far": "utilisé jusqu'à présent",
 		"used.this.time": "utilisé cette fois",
 		"users": "Users\u{26A1}",
+		"user.confirm.delete": "Supprimer l'utilisateur ",
 		"users.failed.to.create.user": "Failed to create a user.\u{26A1}",
 		"users.failed.to.delete.user": "Failed to delete a user.\u{26A1}",
 		"users.failed.to.edit.users": "Failed to edit users.\u{26A1}",
@@ -1142,6 +1146,7 @@ const LocaleTranslationData = {
 		"edit.group": "Edición de grupo",
 		"edit.meter": "Editar Medidor",
 		"edit.unit": "Editar unidad",
+		"edit.user": "Editar Usuario",
 		"email": "Correo Electronico",
 		"enable": "Activar",
 		"error.bar": "Mostrar barras de errores",
@@ -1459,6 +1464,7 @@ const LocaleTranslationData = {
 		"used.so.far": "usado hasta ahora",
 		"used.this.time": "usado esta vez",
 		"users": "Usuarios",
+		"user.delete.confirm": "Borrar usuario ",
 		"users.failed.to.create.user": "No se pudo crear usuario.",
 		"users.failed.to.delete.user": "No se pudo eliminar usuario.",
 		"users.failed.to.edit.users": "No se pudo editar usuarios.",
